requirements.txt
streamlit==1.38.0
python-docx==1.1.2
pdfplumber==0.11.4
pandas==2.2.2
